removed indexjs and copied entire code to server.js in same directory level just for testing
